// html/types.ts
export interface OnDutyUnit {
    identifier: string;
    name: string;
    pos: { x: number, y: number, z: number };
}
// ...
export interface MedicalRecord {
    id: number;
    patient_identifier: string;
    timestamp: string;
    doctor: string;
    diagnosis: string;
    treatment: string;
    notes: string;
}